Add geojson methods in utils
