+++ cartoframes/data/registry/dataframe_dataset.py
@@ -6,10 +6,9 @@ from carto.exceptions import CartoException, CartoRateLimitException
 
 from .base_dataset import BaseDataset
 from ...columns import Column, normalize_names
-from ...geojson import load_geojson
-from ..utils import decode_geometry, compute_geodataframe, detect_encoding_type, is_geojson, \
-    save_index_as_column
-from ...utils import map_geom_type
+from ..utils import decode_geometry, compute_geodataframe, \
+    detect_encoding_type, is_geojson, save_index_as_column
+from ...utils import map_geom_type, load_geojson
 
 
 # avoid _lock issue: https://github.com/tqdm/tqdm/issues/457
