+++ cartoframes/viz/source.py
@@ -3,8 +3,7 @@ from __future__ import absolute_import
 import pandas
 
 from . import defaults
-from ..utils import get_query_bounds
-from ..geojson import get_geodataframe_bounds, encode_geodataframe
+from ..utils import get_query_bounds, get_geodataframe_bounds, encode_geodataframe
 from ..data import Dataset, is_sql_query, is_geojson_file
 
 try:
